requirements.txt
fastapi
uvicorn
torch
torchvision
diffusers
transformers
peft
accelerate
pillow
